[Perf] Export and generalize benchmark Scenario framework
